requirements.txt

python-binance
python-telegram-bot
requests
Flask
Pillow
